Updates to the SES converter
Added a modified version of the US customary units file and the SES converter.  The converter now reads all the way up to the end of the input, generates a pickle file that contains the contents of all the forms and can rewrite an SES input file in SI or US units.  Still to be done is how to handle error 125, cover every variation of the printed output options in the forms and process the output for each timestep.