Added a rant about friction factors
Some people use the symbol 'f' for a friction factor that is one quarter of a friction factor that some other people use the same symbol 'f' for.  This factor of four problem has been around for at least eighty years in the literature.  This is a rant about it.  The document also has some details of what I intend to do about guiding people away from the problem.