Animations of fans in series and fans in parallel
Uploaded .gif files of fans in series interacting and fans in parallel interacting.  These are the same calculation as the .pdf flipbooks uploaded earlier.

If I had had these .gif files when I uploaded the pdf flipbooks of fans in series and fans in parallel, I would have included them.  But at the time ffmpeg was generating .gif files with random yellow fill in the background.  That's now been solved, so here they are.

These .gif files are not as useful as the flipbooks (you have no control over the frame rate or the direction of time's arrow when you look at a .gif) but .gif files are a lot easier for non-IT savvy people to watch.