Uploaded Windows Fortran .exe and .dll files
Replaced "compressible.cp310-win_amd64.pyd" with a recompilation of it and the .dll file that the .pyd file needs to be able to read.

As far as I can tell, having the .pyd file and the .dll file in the same folder as Hobyah.py file means that Hobyah can load the Fortran compressible routines and use them in calculations (but note that it was tested on two Windows 10 machines, both old).