Added zip files of two folders full of data
Added a zip file of the 'old-reference-docs' folder and
a zip file of the 'digitised-test-data' folder.

It turns out that there are a lot of tunnel ventilation engineers who want to get hold of interesting information, but have trouble figuring out how to download files from GitHub in bulk.

This upload adds two zip files: 'digitised-test-data.zip' and
'old-reference-docs.zip'.  These are zip files of the folders of the same name in the Hobyah GitHub repository.

Anyone interested in old tunnel ventilation papers and digitised data from full-scale tests should download these two zip files, unzip them and take a look at README.txt for a description of their contents.